Changes to paste fallback handling
Extracted modal to its own component.  This reduces the code in Toolbar.
It also reduces the responsibilities the modal has.  It is no longer
responsible for making the pasted text available to the VNC session, nor
is it responsible for handing errors in doing so.

There is also no need to clear the textarea. It is unmounted (removed
from the DOM) between each click of "OK" / "Cancel".

The modal copy has also been updated.

The styling of the modal has also been updatd to match the login modal.
There is now an "X" in the top right corner and the "Cancel" button has
changed.